feat(dotfiles): refresh editor sessions and shell helpers
Why
- Improve daily editor startup, project navigation, and shell workflow behavior in the managed dotfiles.
- Keep local sessions and generated artifacts out of tracked files while simplifying secret template entries.

What changed
- Add Neovim session save/load commands with automatic interactive session handling and coordinated nvim-tree startup behavior.
- Add per-filetype indentation and markdown options, explicit Tree-sitter startup, markdown/colorizer plugins, and UI tweaks for diffs, gitsigns, lazygit, and ambiwidth.
- Expand shell helpers with a reload-aware tpd function, a failsafe command runner, and safer nvmrc handling.
- Remove OpenAI/Anthropic entries from the 1Password env template and ignore Session.vim files.

Validation
- Reviewed staged git diff before committing.
- Not run: automated tests.

+++ dot_config/nvim/lua/config/options.lua
@@ -3,7 +3,7 @@ local M = {}
 function M.setup()
 	local opt = vim.opt
 
-	opt.ambiwidth = "double"
+	opt.ambiwidth = "single"
 	opt.autoindent = true
 	opt.autoread = true
 	opt.clipboard = "unnamed"
